[chore]: moved xtasks to modules
